requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.0.3
django-cors-headers==2.2.0
django-nose==1.4.5
django-webpack-loader==0.6.0
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
PyJWT==1.6.0
pytz==2018.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
